+++ src/ansys/fluent/core/launcher/launcher.py
@@ -26,7 +26,7 @@ from ansys.fluent.core.session_meshing import Meshing
 from ansys.fluent.core.session_pure_meshing import PureMeshing
 from ansys.fluent.core.session_solver import Solver
 from ansys.fluent.core.session_solver_icing import SolverIcing
-from ansys.fluent.core.solver.error_message import LauncherError
+from ansys.fluent.core.solver.exceptions import LauncherError
 from ansys.fluent.core.utils.networking import find_remoting_ip
 import ansys.platform.instancemanagement as pypim
