+++ src/ansys/fluent/core/solver/exceptions.py
@@ -69,6 +69,11 @@ class InvalidArgumentError(ValueError):
         super().__init__(error)
 
 
+class InactiveObjectError(RuntimeError):
+    def __init__(self, error):
+        super().__init__(error)
+
+
 class TransactionError:
     """Custom transaction errors."""
