+++ src/ansys/fluent/core/solver/flobject.py
@@ -27,7 +27,7 @@ import sys
 from typing import Any, Dict, Generic, List, NewType, Optional, Tuple, TypeVar, Union
 import weakref
 
-from .error_message import allowed_name_error_message, allowed_values_error
+from .exceptions import allowed_name_error_message, allowed_values_error
 
 settings_logger = logging.getLogger("pyfluent.settings_api")
